requirements.txt
streamlit 
openai
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv